# docs/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: Travio API Gateway
  description: |
    Public-facing REST API for the Travio travel booking ecosystem.
    
    ## Conventions
    - **Casing**: All JSON fields use `snake_case` convention.
    - **Date Format**: `YYYY-MM-DD` for dates, RFC3339 for timestamps.
    
    ## Authentication
    All API requests (except Health, Discovery, and Search) require a valid JWT Access Token.
    Pass the token in the `Authorization` header: `Bearer <token>`.
    
    ## Errors
    Standard error format:
    ```json
    {
      "error": "description of error",
      "code": 400
    }
    ```
  version: 2.0.0
  contact:
    name: Travio Platform Team
    url: https://github.com/MuhibNayem/Travio
servers:
  - url: http://localhost:8888
    description: Local Development Server
...
paths:
...
  /v1/orders:
    post:
      summary: Create a booking order
      description: Initiates the Saga transaction.
      operationId: createOrder
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '202':
          description: Order accepted for processing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateOrderResponse' }
    get:
      summary: List user orders
      operationId: listOrders
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: page_token
          in: query
          schema: { type: string }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListOrdersResponse' }
  /v1/orders/{orderId}:
    get:
      summary: Get order details
      operationId: getOrder
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /v1/orders/{orderId}/cancel:
    post:
      summary: Cancel order
      description: Cancels the order and initiates any necessary refunds.
      operationId: cancelOrder
      tags: [Orders]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Cancellation processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancelOrderResponse' }

  # ================= PAYMENTS =================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [trip_id, from_station_id, to_station_id, passengers, payment_method]
      properties:
        trip_id: { type: string, format: uuid }
        from_station_id: { type: string, format: uuid }
        to_station_id: { type: string, format: uuid }
        hold_id: { type: string, format: uuid, description: "Optional hold ID from Inventory" }
        idempotency_key: { type: string, format: uuid }
        contact_email: { type: string }
        contact_phone: { type: string }
        passengers:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PassengerRequest'
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
    PassengerRequest:
      type: object
      required: [name, seat_id]
      properties:
        nid: { type: string }
        name: { type: string }
        seat_id: { type: string }
        gender: { type: string, enum: [male, female, other] }
        age: { type: integer, minimum: 0 }
        date_of_birth: { type: string }
    CreateOrderResponse:
      type: object
      properties:
        id: { type: string }
        trip_id: { type: string }
        status: { type: string }
        total_paisa: { type: integer }
        # ... other order fields in snake_case
    ListOrdersResponse:
      type: object
      properties:
        orders: 
          type: array
          items: { $ref: '#/components/schemas/Order' }
        total: { type: integer }
        next_page: { type: string }
    Order:
      type: object
      properties:
        id: { type: string }
        status: { type: string }
        total_paisa: { type: integer }
        booking_id: { type: string }
        created_at: { type: string }
        passengers:
          type: array
          items: { $ref: '#/components/schemas/Passenger' }
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, FAILED, CANCELLED, REFUNDED]
    Passenger:
      type: object
      properties:
        name: { type: string }
        seat_number: { type: string }
        seat_class: { type: string }
    CancelOrderResponse:
      type: object
      properties:
        success: { type: boolean }
        order: { $ref: '#/components/schemas/Order' }
        refund: 
            type: object
            properties:
                refund_id: { type: string }
                status: { type: string }

    # === PAYMENT ===
...
    PaymentMethod:
        type: object
        properties:
            id: { type: string }
            name: { type: string }
            enabled: { type: boolean }

    # === FULFILLMENT ===
...